requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
flask-Cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pip==25.0.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
typing_extensions==4.13.2
Werkzeug==3.1.3
